requirements.txt
asn1crypto==0.24.0
boto3==1.9.181
botocore==1.12.181
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
docutils==0.14
ecdsa==0.13.2
Flask==1.0.3
Flask-Login==0.4.1
flask-oidc==1.4.0
flask-talisman==0.7.0
future==0.17.1
httplib2==0.13.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
oauth2client==4.1.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-jose==3.0.1
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4